[CEEZ-10] refactor: 앨범 조회 메서드 albumQueryService로 이동
