+++ src/main/java/com/cheeeese/album/application/AlbumQueryService.java
@@ -1,20 +1,26 @@
 package com.cheeeese.album.application;
 
+import com.cheeeese.album.application.validator.AlbumValidator;
 import com.cheeeese.album.domain.Album;
+import com.cheeeese.album.domain.UserAlbum;
 import com.cheeeese.album.domain.type.Role;
-import com.cheeeese.album.dto.response.ClosedAlbumPageResponse;
-import com.cheeeese.album.dto.response.ClosedAlbumSummaryResponse;
-import com.cheeeese.album.dto.response.OpenAlbumPageResponse;
-import com.cheeeese.album.dto.response.OpenAlbumSummaryResponse;
+import com.cheeeese.album.dto.response.*;
+import com.cheeeese.album.exception.AlbumException;
+import com.cheeeese.album.exception.code.AlbumErrorCode;
+import com.cheeeese.album.infrastructure.mapper.AlbumMapper;
 import com.cheeeese.album.infrastructure.mapper.AlbumQueryMapper;
+import com.cheeeese.album.infrastructure.mapper.UserAlbumMapper;
 import com.cheeeese.album.infrastructure.persistence.UserAlbumRepository;
 import com.cheeeese.cheese4cut.domain.Cheese4cutPhoto;
 import com.cheeeese.cheese4cut.infrastructure.persistence.Cheese4cutPhotoRepository;
+import com.cheeeese.global.security.CustomUserDetails;
 import com.cheeeese.global.util.resolver.CdnUrlResolver;
 import com.cheeeese.photo.domain.Photo;
 import com.cheeeese.photo.domain.PhotoStatus;
+import com.cheeeese.photo.infrastructure.persistence.PhotoLikesRepository;
 import com.cheeeese.photo.infrastructure.persistence.PhotoRepository;
 import com.cheeeese.user.domain.User;
+import com.cheeeese.user.domain.type.ProfileImageType;
 import com.cheeeese.user.exception.UserException;
 import com.cheeeese.user.exception.code.UserErrorCode;
 import com.cheeeese.user.infrastructure.persistence.UserRepository;
@@ -22,6 +28,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -40,8 +48,89 @@ public class AlbumQueryService {
     private final UserAlbumRepository userAlbumRepository;
     private final UserRepository userRepository;
     private final PhotoRepository photoRepository;
+    private final PhotoLikesRepository photoLikesRepository;
     private final Cheese4cutPhotoRepository cheese4cutPhotoRepository;
     private final CdnUrlResolver cdnUrlResolver;
+    private final AlbumValidator albumValidator;
+
+    public UploadAvailableCountResponse getAvailablePhotoCount(String code) {
+        Album album = albumValidator.validateAlbumCode(code);
+
+        int availableCount = album.getRemainingUploadSlots();
+
+        return AlbumMapper.toAvailableCountResponse(
+                availableCount,
+                album.getMaxPhotoCount(),
+                album.getCurrentPhotoCount()
+        );
+    }
+
+    public AlbumParticipantResponse getAlbumParticipantList(Authentication authentication, String code) {
+        User currentUser = extractUser(authentication);
+
+        Album album = albumValidator.validateAlbumCode(code);
+
+        boolean isExpired = album.isExpired();
+
+        Role myRole = null;
+        Long currentUserId = currentUser != null ? currentUser.getId() : null;
+
+        if (currentUserId != null) {
+            Optional<UserAlbum> myUserAlbumOptional = userAlbumRepository.findByUserIdAndAlbumId(currentUserId, album.getId());
+
+            if (myUserAlbumOptional.isPresent()) {
+                myRole = myUserAlbumOptional.get().getRole();
+            }
+        }
+
+        // 앨범의 전체 참여자 목록
+        List<UserAlbum> userAlbums = userAlbumRepository.findAllByAlbumIdExcludeBlack(album.getId(), Role.BLACK);
+
+        List<AlbumParticipantListResponse.ParticipantInfo> participantInfos = buildSortedParticipantInfos(userAlbums, currentUserId);
+
+        return UserAlbumMapper.toAlbumParticipantResponse(
+                album,
+                isExpired,
+                myRole,
+                participantInfos
+        );
+    }
+
+    public AlbumInfoResponse getAlbumInfo(String code) {
+        Album album = albumValidator.validateAlbumCode(code);
+
+        User maker = userAlbumRepository.findMakerByAlbumId(album.getId(), Role.MAKER)
+                .map(UserAlbum::getUser)
+                .orElseThrow(() -> new AlbumException(AlbumErrorCode.USER_NOT_MAKER));
+
+        return AlbumMapper.toAlbumInfoResponse(album, maker);
+    }
+
+    public List<AlbumBest4CutResponse> getAlbumBest4Cut(User user, String code) {
+        Album album = albumValidator.validateAlbumCode(code);
+
+        albumValidator.validateAlbumParticipant(album, user);
+
+        List<Photo> topPhotos = photoRepository.findTop4CompletedPhotosByLikes(
+                album.getId(),
+                PhotoStatus.COMPLETED,
+                PageRequest.of(0, 4)
+        );
+
+        List<Long> photoIds = topPhotos.stream()
+                .map(Photo::getId)
+                .toList();
+
+        Set<Long> likedPhotoIds = photoLikesRepository.findAllLikedPhotoIds(user.getId(), photoIds);
+
+        return topPhotos.stream()
+                .map(photo -> {
+                    String thumbnailUrl = cdnUrlResolver.resolveThumbnail(photo.getThumbnailUrl());
+                    boolean isLiked = likedPhotoIds.contains(photo.getId());
+                    return AlbumMapper.toBest4CutResponse(photo, thumbnailUrl, isLiked);
+                })
+                .toList();
+    }
 
     public OpenAlbumPageResponse getOpenAlbums(User user, int page, int size) {
         Pageable pageable = PageRequest.of(page, size);
@@ -183,5 +272,38 @@ public class AlbumQueryService {
                         Map.Entry::getKey,
                         entry -> List.copyOf(entry.getValue())
                 ));
+    }
+
+    private User extractUser(Authentication authentication) {
+        if (authentication == null || authentication instanceof AnonymousAuthenticationToken) {
+            return null;
+        }
+        Object principal = authentication.getPrincipal();
+        if (principal instanceof CustomUserDetails customUserDetails) {
+            return customUserDetails.getUser();
+        }
+        return null;
+    }
+
+    private List<AlbumParticipantListResponse.ParticipantInfo> buildSortedParticipantInfos(
+            List<UserAlbum> userAlbums,
+            Long currentUserId
+    ) {
+        return userAlbums.stream()
+                .map(userAlbum -> {
+                    User user = userAlbum.getUser();
+                    Role role = userAlbum.getRole();
+                    ProfileImageType type = ProfileImageType.fromName(user.getProfileImage());
+                    String profileImageUrl = cdnUrlResolver.resolveProfile(type.getPath());
+                    boolean isMe = currentUserId != null && user.getId().equals(currentUserId);
+
+                    return UserAlbumMapper.toParticipantInfo(user, profileImageUrl, role, isMe);
+                })
+                .sorted(Comparator
+                        .comparing(AlbumParticipantListResponse.ParticipantInfo::isMe, Comparator.reverseOrder())
+                        .thenComparing(p -> p.role() == Role.MAKER ? 0 : 1)
+                        .thenComparing(AlbumParticipantListResponse.ParticipantInfo::name)
+                )
+                .toList();
     }
 }
