+++ src/main/java/com/cheeeese/album/presentation/AlbumController.java
@@ -1,5 +1,6 @@
 package com.cheeeese.album.presentation;
 
+import com.cheeeese.album.application.AlbumQueryService;
 import com.cheeeese.album.application.AlbumService;
 import com.cheeeese.album.dto.request.AlbumCreationRequest;
 import com.cheeeese.album.dto.response.*;
@@ -11,7 +12,6 @@ import com.cheeeese.user.domain.User;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -24,6 +24,7 @@ import static com.cheeeese.global.common.code.SuccessCode.*;
 public class AlbumController implements AlbumSwagger {
 
     private final AlbumService albumService;
+    private final AlbumQueryService albumQueryService;
 
     @Override
     @PostMapping
@@ -63,7 +64,7 @@ public class AlbumController implements AlbumSwagger {
     @Override
     @GetMapping("/{code}/available-count")
     public CommonResponse<UploadAvailableCountResponse> getAvailableUploadCount(@PathVariable String code) {
-        return CommonResponse.success(PHOTO_AVAILABLE_COUNT_FETCH_SUCCESS, albumService.getAvailablePhotoCount(code));
+        return CommonResponse.success(PHOTO_AVAILABLE_COUNT_FETCH_SUCCESS, albumQueryService.getAvailablePhotoCount(code));
     }
 
     @Override
@@ -74,14 +75,14 @@ public class AlbumController implements AlbumSwagger {
     ) {
       return CommonResponse.success(
               ALBUM_PARTICIPANT_FETCH_SUCCESS,
-              albumService.getAlbumParticipantList(authentication, code)
+              albumQueryService.getAlbumParticipantList(authentication, code)
       );
     }
 
     @Override
     @GetMapping("/{code}/info")
     public CommonResponse<AlbumInfoResponse> getAlbumInfo(@PathVariable String code) {
-        return CommonResponse.success(ALBUM_INFO_GET_SUCCESS, albumService.getAlbumInfo(code));
+        return CommonResponse.success(ALBUM_INFO_GET_SUCCESS, albumQueryService.getAlbumInfo(code));
     }
 
     @Override
@@ -92,7 +93,7 @@ public class AlbumController implements AlbumSwagger {
     ) {
         return CommonResponse.success(
                 ALBUM_BEST4CUT_GET_SUCCESS,
-                albumService.getAlbumBest4Cut(user, code)
+                albumQueryService.getAlbumBest4Cut(user, code)
         );
     }
